Moved WelcomeScreen into pages folder
