+++ personalfinanceapp/src/pages/WelcomeScreen.tsx
